+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/trigger/pojo/PipelineBuildCheckRunStatus.kt
@@ -1,0 +1,4 @@
+package com.tencent.devops.process.trigger.pojo
+
+enum class PipelineBuildCheckRunStatus {
+}
