+++ src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/service/hub/ScmRepositoryApiService.kt
@@ -70,7 +70,6 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import java.time.ZoneId
-import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
 
 /**
